clg bot 2/requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
python-dotenv==1.0.0
nltk==3.8.1
numpy==1.24.3
scikit-learn==1.3.0
python-dateutil==2.8.2
gunicorn==21.2.0
